+++ data/src/main/java/ru/plumsoftware/weatherforecast/data/remote/dto/weatherapi/Location.kt
@@ -8,7 +8,7 @@ data class Location (
   @SerializedName("name"            ) var name           : String? = null,
   @SerializedName("region"          ) var region         : String? = null,
   @SerializedName("country"         ) var country        : String? = null,
-  @SerializedName("lat"             ) var lat            : Int?    = null,
+  @SerializedName("lat"             ) var lat            : Double?    = null,
   @SerializedName("lon"             ) var lon            : Double? = null,
   @SerializedName("tz_id"           ) var tzId           : String? = null,
   @SerializedName("localtime_epoch" ) var localtimeEpoch : Int?    = null,
